chore: refactor getCollectionUrlSlug func
